Improve settings for Order Status and Emails.

+++ src/services/Market_OrderStatusService.php
@@ -136,7 +136,14 @@ class Market_OrderStatusService extends BaseApplicationComponent
      */
     public function deleteById($id)
     {
-        Market_OrderStatusRecord::model()->deleteByPk($id);
+        $statuses = $this->getAll();
+
+        if(count($statuses) >= 2){
+            Market_OrderStatusRecord::model()->deleteByPk($id);
+            return true;
+        }
+
+        return false;
     }
 
     /**
@@ -192,7 +199,6 @@ class Market_OrderStatusService extends BaseApplicationComponent
             $craftEmail->subject = craft()->templates->renderString($email->subject,
                 $renderVariables);
 
-            $body              = $email->type == Market_EmailRecord::TYPE_HTML ? 'htmlBody' : 'body';
             $craftEmail->$body = craft()->templates->render($email->templatePath,
                 $renderVariables);
